refactor(aether-actor): rename the trace_ring module to trace
Module path moves from `aether_actor::trace_ring` to `aether_actor::trace`
for consistency with sibling subsystem naming. Every identifier containing
the substring trace_ring (ActorTraceRing, DEFAULT_TRACE_RING_*, push_trace_ring,
trace_ring_capacity config knobs, test fn names) is unchanged — those name the
ring buffer data structure, not the module.
