+++ crates/aether-substrate/src/config.rs
@@ -81,7 +81,7 @@ use std::fmt;
 use std::fmt::Write as _;
 
 use aether_actor::log::DEFAULT_RING_CAP;
-use aether_actor::trace_ring::{DEFAULT_TRACE_RING_CAP, DEFAULT_TRACE_RING_MAX_CAP};
+use aether_actor::trace::{DEFAULT_TRACE_RING_CAP, DEFAULT_TRACE_RING_MAX_CAP};
 use confique::meta::{Expr, Field, FieldKind, LeafKind, Meta};
 
 use crate::BootError;
@@ -123,7 +123,7 @@ pub struct RingCapacities {
     /// capacity (env `AETHER_ACTOR_LOG_RING_SIZE`; default
     /// [`DEFAULT_RING_CAP`]).
     pub log: usize,
-    /// Per-actor [`ActorTraceRing`](aether_actor::trace_ring::ActorTraceRing)
+    /// Per-actor [`ActorTraceRing`](aether_actor::trace::ActorTraceRing)
     /// and chassis-host-ring *floor* capacity — the size each ring starts
     /// at (env `AETHER_ACTOR_TRACE_RING_SIZE`; default
     /// [`DEFAULT_TRACE_RING_CAP`]).
